Rework internal representation of game data
The main data element is the line now. Lines have a list of other lines
they connect to on both ends. This should make easier to follow a loop
formed by connected lines.

References to other lines, dots and squares are now stored as pointers
instead of id number. This makes the code easier and saner.