penrose: use seed that generates a nice tile
Finally found a seed that generates a roundish tile, like the KrazyDad one.
The seed is 5 fat rhombs forming a star shape, with the star tip pointing
down. This seed only requires 4 foldings to generate the tile.

The side size of the inital rhombs is now properly calculated to produce
a final tile with the correct radius.