solver: fix bottleneck to stop generating artificial solutions
When testing new games for solutions, the bottleneck routine would find
an artificial solution by prolonging a bottleneck for no reason.
It now checks if we have only one loop and all numbered squares have
been handled. If that's the case, stop and give up because otherwise
we'd introduce an artificial solution.