solver: check line before crossing in zero_squares
This function was written under the assumption that
it would only be called only once at the start of the
solution process and, thus, it would not need to check
if the lines were already crossed.

This assumption is no longer valid. The function may
be called more than once. For example, by the
'build_game' routines.