build-game: faster game generation scheme
Instead of put all numbers on the board and take them
out one at a time, we now start with an empty board
and add the numbers one at a time.

The process is faster because at each iteration only
one solution step is needed to know if the number does
something (before the solution had to be carried out
completely). Also, it produces better games.

Needs more work to make the games have the required
difficulty.