gui: convert gui to use actions
Use gtk actions to unify menu items and tool buttons
that perform the same function, i.e., undo, redo ...
A GtkAction is assigned to the menu items and tool button
that receives the events and controls their appearance.

Just discovered that glade is able to save directly
in GtkBuilder format. We don't need to convert from
glade to gtkbuilder.