gui: no more glade, UI is done by hand
Using glade only caused problems and headaches.
We define the GUI by hand using GtkUIManager.

The menu and toolbar are defined in terms of
actions in file 'menu_toolbar.ui'. The actions
are created as GtkActionEntry's in gui.c and
everything is connected by GtkUIManager.

This approach turns out to be easier, shorter
and more robust than using glade.