penrose: optimize speed
Store the center position of a tile when it's created.
This speeds up finding repeated tiles since the center
was computed many times for some tiles (tiles at the end
of the list).

Check the square of the distance. This avoids the
calculation of square roots and speeds up things.