solver: fix bug in bottleneck routine
Assume a situation where we just found one
big open loop, and not all squares have been
handled. Normally, we would then cross out the
connecting line.
However, in the case where this line would
handle all the remaining squares if it were set,
the line can't be crossed out.

Without this test (as was the case before), it is
possible to generate games with more than one
solution.