history: simplify history implementation
Store only one change per history step.
Changes are stored in an array (instead of GList),
this makes traveling through changes as easy as
moving an index. Also when history is rewritten
(i.e. undo and introduce new changes), obsolete
changes don't have to be freed, they'll be overwritten
later as the index goes over.
To be more memory efficient the whole history array
is broken down in segments (of 100 changes) that are
allocated as the history grows. The segments are
non-contiguous in memory and are joined by a GList.