geometry: keep track of clip box needed to redraw correctly
Stop assuming that just the line edges define
the correct clip box for redrawing.
That line change may cause other things to be
drawn also, like a tile's error display.
We keep track of the current clip box inside
the geometry structure and update it when
things are changed and need redrawing.